Objects too big are rescaled

+++ 3DEngine/Source/3DModel.cpp
@@ -81,7 +81,7 @@ void Body3D::SetTransfAngles()
 	angle.x = RadToDeg(angle.x);
 }
 
-void Body3D::SetScale(int new_size)
+void Body3D::SetScale(float new_size)
 {
 	transform.M[0] = transform.M[5] = transform.M[10] = new_size;
 }
