+++ 3DEngine/Source/3DModel.h
@@ -39,7 +39,7 @@ public:
 	
 	// transform
 	void	 SetTransfAngles();
-	void	 SetScale(int new_size);
+	void	 SetScale(float new_size);
 
 	vec3	 GetTransfAngles() const;
 	vec3	 GetTransfScale() const;
